refactor(user): move coursecard to own file
